requirements.txt
beautifulsoup4
requests
selenium
python-dotenv
pymongo[srv]
yfinance
finnhub-python